requirements.txt
torch
transformers
sentence-transformers
langchain
langchain-chroma
langchain-community
chromadb
Flask
huggingface_hub
discord
langchain_huggingface
pandas
python-dotenv
pypdf
matplotlib
openai
huggingface_hub
discord.py
tk
pillow
transformers
sentencepiece
sacremoses
pdfkit
colorama
tabulate